requirements.txt
Flask==1.1.1
line-bot-sdk==1.15.0